% Generated by roxygen2 (4.0.1): do not edit by hand
\name{jekyll.post}
\alias{jekyll.post}
\title{jekyll.post}
\usage{
jekyll.post(filename, title = "", categories = c(), tags = c(),
  layout = F)
}
\arguments{
\item{filename}{The path and name of the RMarkdown file to process.}

\item{title}{The post title}

\item{categories}{A string vector containing the categories this post will belong to}

\item{tags}{A vector of tags for this post}

\item{layout}{The layout to use for this post (defaults to getOption('jekyll.default.template') or 'post')}
}
\description{
Create a post in a Jekyll blog.
}
\details{
jekyll.post `knit`s an RMarkdown file into standard Markdown and saves it
in your Jekyll _posts directory with the appropriate frontmatter prepended.

Expects `options('jekyll.root')` to contain a reference to the root of your
Jekyll blog (the directory that contains `_posts`); putting something like
options(jekyll.root = '~/myblog') in your .Rprofile is probably the easiest
way to do this.
}
\examples{
jekyll.draft('my-post.Rmd', 'My draft', c('R', 'blogs'), c('cats', 'coffee'))
}

